requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3

pydantic-settings==2.6.1
python-multipart==0.0.12

pgvector==0.3.6

sentence-transformers==3.3.1
numpy==2.1.3

httpx==0.27.2

openai==1.57.0
google-genai

boto3==1.35.0
